backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
prance==23.6.21.0
openapi3==1.8.2
swagger-spec-validator==3.0.4
jsonschema==4.23.0
requests==2.32.3
pyyaml==6.0.2
aiofiles==24.1.0
jinja2==3.1.4
faker==33.1.0
websockets==13.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
psutil==6.1.1
PyJWT==2.10.1